Add 'Mount' (analog to 'Capture', etc.)
