+++ servant/servant.cabal
@@ -24,6 +24,7 @@ library
     Servant.API.Delete
     Servant.API.Get
     Servant.API.GetParam
+    Servant.API.Mount
     Servant.API.Post
     Servant.API.Put
     Servant.API.Raw
